Fix some errors for promotion, started carousel admin
Some minor errors from promotion pages
Realised we need an admin page to add, modify and delete carousels.
Thought of an idea for delete promotion, it will be a 2 step process where first you choose and it will redirect you to a 'delete' webpage and you ensure this is the final list. (Thought that it would be good especially if there are a lot of items and you might have accidentally clicked on one)